country was added to the massages
