+++ pyproject.toml
@@ -14,6 +14,7 @@ aiogram = "^3.2.0"
 pendulum = "^3.0.0"
 sqlalchemy = "^2.0.23"
 airportsdata = "^20231017"
+pycountry = "^23.12.11"
 
 [tool.poetry.group.dev.dependencies]
 pytest-asyncio = "^0.23.2"
