+++ tests/telegram_publisher/publisher/test_trip_description.py
@@ -1,5 +1,5 @@
 from telegram_publisher.models import Trip
-from telegram_publisher.publisher import _trip_description
+from telegram_publisher.message_presenter import _trip_description
 
 
 def test_trip_description(trip_first: Trip):
@@ -9,7 +9,6 @@ def test_trip_description(trip_first: Trip):
         '*220 Kč*',
         '🛫 Sun, Jan 1, 04:00 Norvegian Air shuttle',
         '🛬 Mon, Jan 2, 12:00 Norvegian Air shuttle',
-        'approx cost for 1 day\\(s\\):',
         '🏠 3100 Kč   ☕️ 1900 Kč',
         '',
     ]
